requirements.txt
requests
python-dotenv
mysql-connector-python==8.1.0